% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{dire_bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\value{
A string saying hello
}
\description{
dire_bonjour
}
\examples{
dire_bonjour("Denis")
dire_bonjour("")
}
