% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pays.R
\docType{data}
\name{pays}
\alias{pays}
\title{pays}
\format{
A data frame with 241 rows and 6 variables:
\describe{
\item{ X1 }{  numeric ?}
\item{ X2 }{  numeric ?}
\item{ X3 }{  character Abbreviation 2}
\item{ X4 }{  character Abbreviation 3}
\item{ X5 }{  character Name french}
\item{ X6 }{  character Name english}
}
}
\source{
Source
}
\usage{
pays
}
\description{
Description.
}
\keyword{datasets}
